+++ packages/isaacscript-common/src/features/extraConsoleCommands/listCommands.ts
@@ -102,6 +102,7 @@ import {
   listGridEntities,
   movePlayer,
   spawnTrapdoorOrCrawlSpace,
+  warpNextToRoomType,
   warpToRoomType,
 } from "./commandsSubroutines";
 import v from "./v";
@@ -282,6 +283,11 @@ export function bombs(params: string): void {
  */
 export function boneHearts(params: string): void {
   addHeart(params, HealthType.BONE);
+}
+
+/** Warps to the room next to the first Boss Room on the floor. */
+export function bossNextRoom(): void {
+  warpNextToRoomType(RoomType.BOSS);
 }
 
 /** Warps to the first Boss Room on the floor. */