+++ packages/isaacscript-common/src/objects/roomTypeNames.ts
@@ -1,7 +1,5 @@
 import { RoomType } from "isaac-typescript-definitions";
 
-export const DEFAULT_ROOM_TYPE_NAME = "Unknown";
-
 export const ROOM_TYPE_NAMES: { readonly [key in RoomType]: string } = {
   [RoomType.DEFAULT]: "default room", // 1
   [RoomType.SHOP]: "shop", // 2